requirements.txt
asgiref==3.11.1
asttokens==3.0.1
decorator==5.2.1
Django==5.2.8
django-cid==3.0
django-cors-headers==4.9.0
django-environ==0.13.0
django-extensions==4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
executing==2.2.1
ipython==9.11.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.2.1
parso==0.8.6
pexpect==4.9.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.12.1
python-ulid==3.1.0
pytz==2026.1.post1
sqlparse==0.5.5
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.15.0
wcwidth==0.6.0
